requirements.txt
instagrapi==2.0.0
click==8.1.7
colorama==0.4.6
python-dotenv==1.0.0
keyring==24.3.0
Pillow>=8.1.1